+++ activerecord/test/cases/arel/visitors/dot_test.rb
@@ -85,6 +85,91 @@ module Arel
         collector = Collectors::PlainString.new
         assert_match '[label="<f0>ActiveModel::Attribute::WithCastValue"]', @visitor.accept(node, collector).value
       end
+
+      def test_Arel_Nodes_CurrentRow
+        node = Arel::Nodes::CurrentRow.new
+        collector = Collectors::PlainString.new
+        assert_match '[label="<f0>Arel::Nodes::CurrentRow"]', @visitor.accept(node, collector).value
+      end
+
+      def test_Arel_Nodes_Distinct
+        node = Arel::Nodes::Distinct.new
+        collector = Collectors::PlainString.new
+        assert_match '[label="<f0>Arel::Nodes::Distinct"]', @visitor.accept(node, collector).value
+      end
+
+      def test_Arel_Nodes_Case_and_friends
+        foo = Arel::Nodes.build_quoted("foo")
+        node = Arel::Nodes::Case.new(foo)
+        node.conditions = [Arel::Nodes::When.new(foo, Arel::Nodes.build_quoted(1))]
+        node.default = Arel::Nodes::Else.new(Arel::Nodes.build_quoted(0))
+
+        dot = @visitor.accept(node, Arel::Collectors::PlainString.new).value
+
+        assert_match '[label="<f0>Arel::Nodes::Case"]', dot
+        assert_edge("case", dot)
+        assert_edge("conditions", dot)
+        assert_edge("default", dot)
+        assert_match '[label="<f0>Arel::Nodes::When"]', dot
+        assert_match '[label="<f0>Arel::Nodes::Else"]', dot
+        assert_match '[label="<f0>Arel::Nodes::Else"]', dot
+      end
+
+      def test_Arel_Nodes_InfixOperation
+        node = Arel::Nodes::InfixOperation.new("&&", Arel::Nodes.build_quoted(1), Arel::Nodes.build_quoted(2))
+
+        dot = @visitor.accept(node, Arel::Collectors::PlainString.new).value
+
+        assert_match '[label="<f0>Arel::Nodes::InfixOperation"]', dot
+        assert_edge("operator", dot)
+        assert_edge("left", dot)
+        assert_edge("right", dot)
+      end
+
+      def test_Arel_Nodes_RegExp
+        table = Table.new(:users)
+        node = Arel::Nodes::Regexp.new(table[:name], Nodes.build_quoted("foo%"))
+
+        dot = @visitor.accept(node, Arel::Collectors::PlainString.new).value
+
+        assert_match '[label="<f0>Arel::Nodes::Regexp"]', dot
+        assert_edge("left", dot)
+        assert_edge("right", dot)
+        assert_edge("case_sensitive", dot)
+      end
+
+      def test_Arel_Nodes_NotRegExp
+        table = Table.new(:users)
+        node = Arel::Nodes::NotRegexp.new(table[:name], Nodes.build_quoted("foo%"))
+
+        dot = @visitor.accept(node, Arel::Collectors::PlainString.new).value
+
+        assert_match '[label="<f0>Arel::Nodes::NotRegexp"]', dot
+        assert_edge("left", dot)
+        assert_edge("right", dot)
+        assert_edge("case_sensitive", dot)
+      end
+
+      def test_Arel_Nodes_UnaryOperation
+        node = Arel::Nodes::UnaryOperation.new(:-, 1)
+
+        dot = @visitor.accept(node, Arel::Collectors::PlainString.new).value
+
+        assert_match '[label="<f0>Arel::Nodes::UnaryOperation"]', dot
+        assert_edge("operator", dot)
+        assert_edge("expr", dot)
+      end
+
+      def test_Arel_Nodes_With
+        node = Arel::Nodes::With.new(["query1", "query2", "query3"])
+
+        dot = @visitor.accept(node, Arel::Collectors::PlainString.new).value
+
+        assert_match '[label="<f0>Arel::Nodes::With"]', dot
+        assert_edge("0", dot)
+        assert_edge("1", dot)
+        assert_edge("2", dot)
+      end
     end
   end
 end